Adicionado acesso a API através de POST
-- Acessar /app usando POST manda o request diretamente à API
-- Refatorado código, classes HttpErros e HttpFile agora são Http, os métodos de ambas foram mescladas nessa nova classe
-- Adicionado método getPOST() na classe Http, que recebe o requestBody de input e retorna uma String do POST
-- Adicionado classe package api com classe Input para ser o ponto de entrada de processamento de todas as requests à API
