+++ src/api/Input.java
@@ -1,0 +1,7 @@
+package api;
+
+public class Input {
+	public static void process(String query) {
+		System.out.println("'"+query+"'");
+	}
+}
